+++ Points.h
@@ -3,12 +3,12 @@
 class Points
 {
 public:
-	Points(uint16_t x, uint16_t y, uint8_t R, uint8_t G, uint8_t B, uint8_t brushSize, bool inWindow);
+	Points(uint16_t x, uint16_t y, uint16_t R, uint16_t G, uint16_t B, uint8_t brushSize, bool inWindow);
 	const uint16_t GetX() const;
 	const uint16_t GetY() const;
-	const uint8_t GetRed() const;
-	const uint8_t GetGreen() const;
-	const uint8_t GetBlue() const;
+	const uint16_t GetRed() const;
+	const uint16_t GetGreen() const;
+	const uint16_t GetBlue() const;
 	const uint8_t GetBrushSize() const;
 	const bool GetInWindow() const;
 	void SetX(uint16_t x);
@@ -23,9 +23,9 @@ public:
 private:
 	uint16_t m_x;
 	uint16_t m_y;
-	uint8_t m_R;
-	uint8_t m_G;
-	uint8_t m_B;
+	uint16_t m_R;
+	uint16_t m_G;
+	uint16_t m_B;
 	uint8_t m_brushSize;
 	bool m_inWindow;
